requirements.txt
Django==1.9.6
django-debug-toolbar==1.4
django-extensions==1.6.7
django-filter==0.13.0
djangorestframework==3.3
requests-cache==0.4.12
